Rename wiringcmd plugin to cmdbuilder to make it consistent with documentation
