+++ examples/dsb_sn/wiring/main.go
@@ -3,12 +3,12 @@
 //
 // To display options and usage, invoke:
 //
-//  go run main.go -h
+//	go run main.go -h
 package main
 
 import (
 	"github.com/blueprint-uservices/blueprint/examples/dsb_sn/wiring/specs"
-	"github.com/blueprint-uservices/blueprint/plugins/wiringcmd"
+	"github.com/blueprint-uservices/blueprint/plugins/cmdbuilder"
 	"github.com/blueprint-uservices/blueprint/plugins/workflow"
 )
 
@@ -18,7 +18,7 @@ func main() {
 
 	// Build a supported wiring spec
 	name := "SocialNetwork"
-	wiringcmd.MakeAndExecute(
+	cmdbuilder.MakeAndExecute(
 		name,
 		specs.Docker,
 	)
