+++ examples/dsb_sn/wiring/specs/README.md
@@ -1,0 +1,26 @@
+<!-- Code generated by gomarkdoc. DO NOT EDIT -->
+
+# specs
+
+```go
+import "github.com/blueprint-uservices/blueprint/examples/dsb_sn/wiring/specs"
+```
+
+## Index
+
+- [Variables](<#variables>)
+
+
+## Variables
+
+<a name="Docker"></a>A wiring spec that deploys each service into its own Docker container and using thrift to communicate between services. All services except the Wrk2API service use thrift for communication; WRK2API service provides the http frontend. The user, socialgraph, urlshorten, and usertimeline services use MongoDB instances to store their data. The user, socialgraph, urlshorten, usertimeine, and hometimeline services use memcached instances as the cache data for faster responses.
+
+```go
+var Docker = cmdbuilder.SpecOption{
+    Name:        "docker",
+    Description: "Deploys each service in a separate container with thrift, and uses mongodb as NoSQL database backends.",
+    Build:       makeDockerSpec,
+}
+```
+
+Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
