+++ examples/leaf/wiring/README.md
@@ -1,10 +1,17 @@
-# Leaf Application - Wiring Spec
+<!-- Code generated by gomarkdoc. DO NOT EDIT -->
 
-`main.go` defines the wiring spec for the Leaf application.  It instantiates the leaf and nonleaf services, applies some modifiers, and deploys them in separate processes with GRPC for communication
+# wiring
 
-From this directory, run
+```go
+import "github.com/blueprint-uservices/blueprint/examples/leaf/wiring"
 ```
-go run main.go
-```
 
-Doing so will print out the IR of the application.
+Package main provides the LeafApp application, primarily used for testing compilation during the plugin development process.
+
+Run with go run examples/leaf/wiring/main.go
+
+## Index
+
+
+
+Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
