+++ examples/leaf/wiring/main.go
@@ -6,7 +6,7 @@ package main
 
 import (
 	"github.com/blueprint-uservices/blueprint/examples/leaf/wiring/specs"
-	"github.com/blueprint-uservices/blueprint/plugins/wiringcmd"
+	"github.com/blueprint-uservices/blueprint/plugins/cmdbuilder"
 	"github.com/blueprint-uservices/blueprint/plugins/workflow"
 )
 
@@ -16,7 +16,7 @@ func main() {
 
 	// Build a supported wiring spec
 	name := "LeafApp"
-	wiringcmd.MakeAndExecute(
+	cmdbuilder.MakeAndExecute(
 		name,
 		specs.Docker,
 		specs.Thrift,
