+++ examples/leaf/wiring/specs/README.md
@@ -1,0 +1,96 @@
+<!-- Code generated by gomarkdoc. DO NOT EDIT -->
+
+# specs
+
+```go
+import "github.com/blueprint-uservices/blueprint/examples/leaf/wiring/specs"
+```
+
+## Index
+
+- [Variables](<#variables>)
+
+
+## Variables
+
+<a name="Docker"></a>
+
+```go
+var Docker = cmdbuilder.SpecOption{
+    Name:        "docker",
+    Description: "Deploys each service in a separate container with gRPC, uses mongodb as NoSQL database backends, and applies a number of modifiers.",
+    Build:       makeDockerSpec,
+}
+```
+
+<a name="Govector"></a>Wiring specification demonstrates how to instrument an application using GoVector to propagate vector clocks and to create logs with vector clocks. The wiring spec uses govector logger as the custom logger for processes. Each service is deployed in a separate process with services communicating using HTTP.
+
+```go
+var Govector = cmdbuilder.SpecOption{
+    Name:        "govector",
+    Description: "Deploys each service in a separate process, communicating using HTTP. Wraps each service in GoVector vector clocks and sets the GoVectorLogger for each process.",
+    Build:       makeGoVectorLoggerSpec,
+}
+```
+
+<a name="HTTP"></a>
+
+```go
+var HTTP = cmdbuilder.SpecOption{
+    Name:        "http",
+    Description: "Deploys each service in a separate process, communicating using HTTP.  Wraps each service in Zipkin tracing.",
+    Build:       makeHTTPSpec,
+}
+```
+
+<a name="OT_Logger"></a>Wiring specification demonstrates how to use a custom logger for Blueprint applications. The wiring spec uses opentelemetry logger as the custom logger for demonstration. Each service is deployed in a separate container with services communicating using HTTP. Launches an zipkin server which collects spans generated from each service wrapped in opentelemetry tracing.
+
+```go
+var OT_Logger = cmdbuilder.SpecOption{
+    Name:        "ot_logger",
+    Description: "Deploys each service in a separate container, communicating using HTTP. Wraps each service in opentelemetry tracing and sets the OTLogger for each process. All spans are collected by the zipkin collector.",
+    Build:       makeOTLoggerSpec,
+}
+```
+
+<a name="Thrift"></a>
+
+```go
+var Thrift = cmdbuilder.SpecOption{
+    Name:        "thrift",
+    Description: "Deploys each service in a separate process, communicating using Thrift.",
+    Build:       makeThriftSpec,
+}
+```
+
+<a name="TimeoutDemo"></a>A wiring spec that demonstrates how to add timeouts to a blueprint application. The spec deploys each service in a separate container. The services use GRPC to communicate with each other. Server side of each service is configured with a latency injector which adds a fixed amount of latency for every request. Client side for each service is configured with timeouts. All requests in the generated system with this wiring specification result in a TimeOut error.
+
+```go
+var TimeoutDemo = cmdbuilder.SpecOption{
+    Name:        "timeout_demo",
+    Description: "Deploys each service in a separate container with gRPC and configures the clients with timeouts and the servers with latency injectors to demonstrate timeouts in blueprint",
+    Build:       makeDockerTimeoutSpec,
+}
+```
+
+<a name="TimeoutRetriesDemo"></a>A wiring spec that demonstrates how to add timeouts with retries to a blueprint application. The spec deploys each service in a separate container. The services use GRPC to communicate with each other. Server side of each service is configured with a latency injector which adds a fixed amount of latency for every request. Client side for each service is configured with retries where each separate request results in a timeout. All requests in the generated system with this wiring specification result in a TimeOut error.
+
+```go
+var TimeoutRetriesDemo = cmdbuilder.SpecOption{
+    Name:        "timeout_retries_demo",
+    Description: "Deploys each service in a separate container with gRPC and configures the clients with both retries and timeouts and the servers with latency injectors to demonstrate timeouts in blueprint",
+    Build:       makeDockerTimeoutRetriesSpec,
+}
+```
+
+<a name="Xtrace_Logger"></a>Wiring specification demonstrates how to use a custom logger for Blueprint applications. The wiring spec uses xtrace logger as the custom logger for demonstration. Each service is deployed in a separate container with services communicating using HTTP. Launches an XTrace server with each service wrapped in xtrace tracing.
+
+```go
+var Xtrace_Logger = cmdbuilder.SpecOption{
+    Name:        "xtrace_logger",
+    Description: "Deploys each service in a separate container, communicating using HTTP. Wraps each service in XTrace tracing and sets the XTraceLogger for each process.",
+    Build:       makeXTraceLoggerSpec,
+}
+```
+
+Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
