+++ examples/leaf/wiring/specs/govector.go
@@ -4,18 +4,18 @@ import (
 	"strings"
 
 	"github.com/blueprint-uservices/blueprint/blueprint/pkg/wiring"
+	"github.com/blueprint-uservices/blueprint/plugins/cmdbuilder"
 	"github.com/blueprint-uservices/blueprint/plugins/goproc"
 	"github.com/blueprint-uservices/blueprint/plugins/govector"
 	"github.com/blueprint-uservices/blueprint/plugins/http"
 	"github.com/blueprint-uservices/blueprint/plugins/simple"
-	"github.com/blueprint-uservices/blueprint/plugins/wiringcmd"
 	"github.com/blueprint-uservices/blueprint/plugins/workflow"
 )
 
 // Wiring specification demonstrates how to instrument an application using GoVector to propagate vector clocks and to create logs with vector clocks.
 // The wiring spec uses govector logger as the custom logger for processes.
 // Each service is deployed in a separate process with services communicating using HTTP.
-var Govector = wiringcmd.SpecOption{
+var Govector = cmdbuilder.SpecOption{
 	Name:        "govector",
 	Description: "Deploys each service in a separate process, communicating using HTTP. Wraps each service in GoVector vector clocks and sets the GoVectorLogger for each process.",
 	Build:       makeGoVectorLoggerSpec,
