+++ examples/leaf/wiring/specs/timeout_demo.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/blueprint-uservices/blueprint/blueprint/pkg/wiring"
+	"github.com/blueprint-uservices/blueprint/plugins/cmdbuilder"
 	"github.com/blueprint-uservices/blueprint/plugins/goproc"
 	"github.com/blueprint-uservices/blueprint/plugins/http"
 	"github.com/blueprint-uservices/blueprint/plugins/latency"
@@ -12,7 +13,6 @@ import (
 	"github.com/blueprint-uservices/blueprint/plugins/retries"
 	"github.com/blueprint-uservices/blueprint/plugins/simple"
 	"github.com/blueprint-uservices/blueprint/plugins/timeouts"
-	"github.com/blueprint-uservices/blueprint/plugins/wiringcmd"
 	"github.com/blueprint-uservices/blueprint/plugins/workflow"
 )
 
@@ -22,7 +22,7 @@ import (
 // Server side of each service is configured with a latency injector which adds a fixed amount of latency for every request.
 // Client side for each service is configured with timeouts.
 // All requests in the generated system with this wiring specification result in a TimeOut error.
-var TimeoutDemo = wiringcmd.SpecOption{
+var TimeoutDemo = cmdbuilder.SpecOption{
 	Name:        "timeout_demo",
 	Description: "Deploys each service in a separate container with gRPC and configures the clients with timeouts and the servers with latency injectors to demonstrate timeouts in blueprint",
 	Build:       makeDockerTimeoutSpec,
@@ -34,7 +34,7 @@ var TimeoutDemo = wiringcmd.SpecOption{
 // Server side of each service is configured with a latency injector which adds a fixed amount of latency for every request.
 // Client side for each service is configured with retries where each separate request results in a timeout.
 // All requests in the generated system with this wiring specification result in a TimeOut error.
-var TimeoutRetriesDemo = wiringcmd.SpecOption{
+var TimeoutRetriesDemo = cmdbuilder.SpecOption{
 	Name:        "timeout_retries_demo",
 	Description: "Deploys each service in a separate container with gRPC and configures the clients with both retries and timeouts and the servers with latency injectors to demonstrate timeouts in blueprint",
 	Build:       makeDockerTimeoutRetriesSpec,
