+++ examples/leaf/workflow/leaf/README.md
@@ -1,0 +1,188 @@
+<!-- Code generated by gomarkdoc. DO NOT EDIT -->
+
+# leaf
+
+```go
+import "github.com/blueprint-uservices/blueprint/examples/leaf/workflow/leaf"
+```
+
+## Index
+
+- [type LeafObject](<#LeafObject>)
+- [type LeafService](<#LeafService>)
+- [type LeafServiceImpl](<#LeafServiceImpl>)
+  - [func NewLeafServiceImpl\(ctx ctxx.Context, cache backend.Cache, db backend.NoSQLDatabase\) \(\*LeafServiceImpl, error\)](<#NewLeafServiceImpl>)
+  - [func \(l \*LeafServiceImpl\) HelloInt\(ctx ctxx.Context, a int64\) \(int64, error\)](<#LeafServiceImpl.HelloInt>)
+  - [func \(l \*LeafServiceImpl\) HelloMate\(ctx ctxx.Context, a int, b int32, c string, d map\[string\]LeafObject, elems \[\]string, elems2 \[\]NestedLeafObject\) \(string, \[\]string, int32, int, map\[string\]LeafObject, error\)](<#LeafServiceImpl.HelloMate>)
+  - [func \(l \*LeafServiceImpl\) HelloNothing\(ctx ctxx.Context\) error](<#LeafServiceImpl.HelloNothing>)
+  - [func \(l \*LeafServiceImpl\) HelloObject\(ctx ctxx.Context, obj \*LeafObject\) \(\*LeafObject, error\)](<#LeafServiceImpl.HelloObject>)
+  - [func \(l \*LeafServiceImpl\) NonServiceFunction\(\) int64](<#LeafServiceImpl.NonServiceFunction>)
+- [type MyInt](<#MyInt>)
+- [type NestedLeafObject](<#NestedLeafObject>)
+- [type NonLeafService](<#NonLeafService>)
+  - [func NewNonLeafServiceImpl\(ctx context.Context, leafService LeafService\) \(NonLeafService, error\)](<#NewNonLeafServiceImpl>)
+- [type NonLeafServiceImpl](<#NonLeafServiceImpl>)
+  - [func \(nl \*NonLeafServiceImpl\) Hello\(ctx context.Context, a int64\) \(int64, error\)](<#NonLeafServiceImpl.Hello>)
+
+
+<a name="LeafObject"></a>
+## type [LeafObject](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/leafservice.go#L20-L25>)
+
+
+
+```go
+type LeafObject struct {
+    ID    int64
+    Name  string
+    Count int
+    Props map[string]NestedLeafObject
+}
+```
+
+<a name="LeafService"></a>
+## type [LeafService](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/leafservice.go#L27-L32>)
+
+
+
+```go
+type LeafService interface {
+    HelloNothing(ctx ctxx.Context) error
+    HelloInt(ctx ctxx.Context, a int64) (int64, error)
+    HelloObject(ctx ctxx.Context, obj *LeafObject) (*LeafObject, error)
+    HelloMate(ctx ctxx.Context, a int, b int32, c string, d map[string]LeafObject, elems []string, elems2 []NestedLeafObject) (string, []string, int32, int, map[string]LeafObject, error)
+}
+```
+
+<a name="LeafServiceImpl"></a>
+## type [LeafServiceImpl](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/leafservice.go#L34-L40>)
+
+
+
+```go
+type LeafServiceImpl struct {
+    LeafService
+    Cache      backend.Cache
+    Collection backend.NoSQLCollection
+    Counter    metric.Int64Counter
+    // contains filtered or unexported fields
+}
+```
+
+<a name="NewLeafServiceImpl"></a>
+### func [NewLeafServiceImpl](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/leafservice.go#L42>)
+
+```go
+func NewLeafServiceImpl(ctx ctxx.Context, cache backend.Cache, db backend.NoSQLDatabase) (*LeafServiceImpl, error)
+```
+
+
+
+<a name="LeafServiceImpl.HelloInt"></a>
+### func \(\*LeafServiceImpl\) [HelloInt](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/leafservice.go#L65>)
+
+```go
+func (l *LeafServiceImpl) HelloInt(ctx ctxx.Context, a int64) (int64, error)
+```
+
+
+
+<a name="LeafServiceImpl.HelloMate"></a>
+### func \(\*LeafServiceImpl\) [HelloMate](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/leafservice.go#L105>)
+
+```go
+func (l *LeafServiceImpl) HelloMate(ctx ctxx.Context, a int, b int32, c string, d map[string]LeafObject, elems []string, elems2 []NestedLeafObject) (string, []string, int32, int, map[string]LeafObject, error)
+```
+
+
+
+<a name="LeafServiceImpl.HelloNothing"></a>
+### func \(\*LeafServiceImpl\) [HelloNothing](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/leafservice.go#L59>)
+
+```go
+func (l *LeafServiceImpl) HelloNothing(ctx ctxx.Context) error
+```
+
+
+
+<a name="LeafServiceImpl.HelloObject"></a>
+### func \(\*LeafServiceImpl\) [HelloObject](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/leafservice.go#L100>)
+
+```go
+func (l *LeafServiceImpl) HelloObject(ctx ctxx.Context, obj *LeafObject) (*LeafObject, error)
+```
+
+
+
+<a name="LeafServiceImpl.NonServiceFunction"></a>
+### func \(\*LeafServiceImpl\) [NonServiceFunction](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/leafservice.go#L110>)
+
+```go
+func (l *LeafServiceImpl) NonServiceFunction() int64
+```
+
+
+
+<a name="MyInt"></a>
+## type [MyInt](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/leafservice.go#L12>)
+
+
+
+```go
+type MyInt int64
+```
+
+<a name="NestedLeafObject"></a>
+## type [NestedLeafObject](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/leafservice.go#L14-L18>)
+
+
+
+```go
+type NestedLeafObject struct {
+    Key   string
+    Value string
+    Props []string
+}
+```
+
+<a name="NonLeafService"></a>
+## type [NonLeafService](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/nonleafservice.go#L10-L12>)
+
+
+
+```go
+type NonLeafService interface {
+    Hello(ctx context.Context, a int64) (int64, error)
+}
+```
+
+<a name="NewNonLeafServiceImpl"></a>
+### func [NewNonLeafServiceImpl](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/nonleafservice.go#L20>)
+
+```go
+func NewNonLeafServiceImpl(ctx context.Context, leafService LeafService) (NonLeafService, error)
+```
+
+
+
+<a name="NonLeafServiceImpl"></a>
+## type [NonLeafServiceImpl](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/nonleafservice.go#L14-L18>)
+
+
+
+```go
+type NonLeafServiceImpl struct {
+    NonLeafService
+    // contains filtered or unexported fields
+}
+```
+
+<a name="NonLeafServiceImpl.Hello"></a>
+### func \(\*NonLeafServiceImpl\) [Hello](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/leaf/workflow/leaf/nonleafservice.go#L27>)
+
+```go
+func (nl *NonLeafServiceImpl) Hello(ctx context.Context, a int64) (int64, error)
+```
+
+
+
+Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
