+++ examples/sockshop/wiring/main.go
@@ -9,7 +9,7 @@ package main
 
 import (
 	"github.com/blueprint-uservices/blueprint/examples/sockshop/wiring/specs"
-	"github.com/blueprint-uservices/blueprint/plugins/wiringcmd"
+	"github.com/blueprint-uservices/blueprint/plugins/cmdbuilder"
 	"github.com/blueprint-uservices/blueprint/plugins/workflow"
 )
 
@@ -19,7 +19,7 @@ func main() {
 
 	// Build a supported wiring spec
 	name := "SockShop"
-	wiringcmd.MakeAndExecute(
+	cmdbuilder.MakeAndExecute(
 		name,
 		specs.Basic,
 		specs.GRPC,
