+++ examples/train_ticket/wiring/README.md
@@ -1,0 +1,21 @@
+<!-- Code generated by gomarkdoc. DO NOT EDIT -->
+
+# wiring
+
+```go
+import "github.com/blueprint-uservices/blueprint/examples/train_ticket/wiring"
+```
+
+Package main provides an application for compiling different wiring specs for TrainTicket application.
+
+To display options and usage, invoke:
+
+```
+go run main.go -h
+```
+
+## Index
+
+
+
+Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
