+++ examples/train_ticket/wiring/specs/README.md
@@ -1,0 +1,26 @@
+<!-- Code generated by gomarkdoc. DO NOT EDIT -->
+
+# specs
+
+```go
+import "github.com/blueprint-uservices/blueprint/examples/train_ticket/wiring/specs"
+```
+
+## Index
+
+- [Variables](<#variables>)
+
+
+## Variables
+
+<a name="Docker"></a>A wiring spec that deploys each service into its own Docker container and uses http to communicate between services. The user service uses MongoDB instance to store their data.
+
+```go
+var Docker = cmdbuilder.SpecOption{
+    Name:        "docker",
+    Description: "Deploys each service in a separate container with http, and uses mongodb as NoSQL database backends",
+    Build:       makeDockerSpec,
+}
+```
+
+Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
