+++ examples/train_ticket/workflow/food/README.md
@@ -13,7 +13,7 @@ import "github.com/blueprint-uservices/blueprint/examples/train_ticket/workflow/
 
 
 <a name="Food"></a>
-## type [Food](<https://github.com/Blueprint-uServices/blueprint/blob/main/examples/train_ticket/workflow/food/data.go#L3-L6>)
+## type [Food](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/train_ticket/workflow/food/data.go#L3-L6>)
 
 
 
@@ -25,7 +25,7 @@ type Food struct {
 ```
 
 <a name="FoodOrder"></a>
-## type [FoodOrder](<https://github.com/Blueprint-uServices/blueprint/blob/main/examples/train_ticket/workflow/food/data.go#L8-L16>)
+## type [FoodOrder](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/train_ticket/workflow/food/data.go#L8-L16>)
