+++ examples/train_ticket/workflow/news/README.md
@@ -18,7 +18,7 @@ package news implements the ts\-news\-service from the TrainTicket application
 
 
 <a name="News"></a>
-## type [News](<https://github.com/Blueprint-uServices/blueprint/blob/main/examples/train_ticket/workflow/news/data.go#L3-L6>)
+## type [News](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/train_ticket/workflow/news/data.go#L3-L6>)
 
 
 
@@ -30,7 +30,7 @@ type News struct {
 ```
 
 <a name="NewsService"></a>
-## type [NewsService](<https://github.com/Blueprint-uServices/blueprint/blob/main/examples/train_ticket/workflow/news/newsService.go#L7-L9>)
+## type [NewsService](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/train_ticket/workflow/news/newsService.go#L7-L9>)
 
 News Service provides the latest news about the application
 
@@ -41,7 +41,7 @@ type NewsService interface {
 ```
 
 <a name="NewsServiceImpl"></a>
-## type [NewsServiceImpl](<https://github.com/Blueprint-uServices/blueprint/blob/main/examples/train_ticket/workflow/news/newsService.go#L12>)
+## type [NewsServiceImpl](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/train_ticket/workflow/news/newsService.go#L12>)
 
 News Service Implementation
 
@@ -50,7 +50,7 @@ type NewsServiceImpl struct{}
 ```
 
 <a name="NewNewsServiceImpl"></a>
-### func [NewNewsServiceImpl](<https://github.com/Blueprint-uServices/blueprint/blob/main/examples/train_ticket/workflow/news/newsService.go#L14>)
+### func [NewNewsServiceImpl](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/train_ticket/workflow/news/newsService.go#L14>)
 
 ```go
 func NewNewsServiceImpl(ctx context.Context) (*NewsServiceImpl, error)
@@ -59,7 +59,7 @@ func NewNewsServiceImpl(ctx context.Context) (*NewsServiceImpl, error)
 
 
 <a name="NewsServiceImpl.Hello"></a>
-### func \(\*NewsServiceImpl\) [Hello](<https://github.com/Blueprint-uServices/blueprint/blob/main/examples/train_ticket/workflow/news/newsService.go#L18>)
+### func \(\*NewsServiceImpl\) [Hello](<https://github.com/blueprint-uservices/blueprint/blob/main/examples/train_ticket/workflow/news/newsService.go#L18>)
 
 ```go
 func (n *NewsServiceImpl) Hello(ctx context.Context, val string) (string, error)
