+++ plugins/cmdbuilder/README.md
@@ -1,12 +1,12 @@
 <!-- Code generated by gomarkdoc. DO NOT EDIT -->
 
-# wiringcmd
+# cmdbuilder
 
 ```go
-import "github.com/blueprint-uservices/blueprint/plugins/wiringcmd"
+import "github.com/blueprint-uservices/blueprint/plugins/cmdbuilder"
 ```
 
-Package wiringcmd is a helper package for building wiring spec command line programs. It doesn't provide any wiring spec commands or IR.
+Package cmdbuilder is a helper package for building wiring spec command line programs. It doesn't provide any wiring spec commands or IR.
 
 The [CmdBuilder](<#CmdBuilder>) struct enables an application to register multiple wiring spec options in a single main.go. It adds command line arguments for selecting which wiring spec to compile, and takes care of argument parsing and spec building.
 
@@ -29,7 +29,7 @@ func buildMySpec(spec wiring.WiringSpec) ([]string, error) {
 	... // does wiring spec stuff
 }
 
-var MySpec = wiringcmd.SpecOption{
+var MySpec = cmdbuilder.SpecOption{
 	Name:        "myspec",
 	Description: "My example wiring spec",
 	Build:       buildMySpec,
@@ -39,13 +39,13 @@ var MySpec = wiringcmd.SpecOption{
 Lastly, in the main file, call [MakeAndExecute](<#MakeAndExecute>):
 
 ```
-wiringcmd.MakeAndExecute(
+cmdbuilder.MakeAndExecute(
 	"MyApplication",
 	MySpec
 )
 ```
 
-[MakeAndExecute](<#MakeAndExecute>) accepts any number of specs. wiringcmd takes care of parsing command line args
+[MakeAndExecute](<#MakeAndExecute>) accepts any number of specs. cmdbuilder takes care of parsing command line args
 
 ### Runtime Usage
 
@@ -55,7 +55,7 @@ Run your program with
 go run main.go -h
 ```
 
-The wiringcmd plugin takes care of argument parsing, and will list the wiring specs that can be compiled.
+The cmdbuilder plugin takes care of argument parsing, and will list the wiring specs that can be compiled.
 
 To compile a spec, run
 
@@ -77,7 +77,7 @@ go run main.go -o build -w myspec
 
 
 <a name="MakeAndExecute"></a>
-## func [MakeAndExecute](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/wiringcmd/cmdbuilder.go#L100>)
+## func [MakeAndExecute](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/cmdbuilder/cmdbuilder.go#L100>)
 
 ```go
 func MakeAndExecute(name string, specs ...SpecOption)
@@ -86,7 +86,7 @@ func MakeAndExecute(name string, specs ...SpecOption)
 Parses command line flags, and if a valid spec is specified with the \-w flag, that exists within specs, executes that spec.
 
 <a name="CmdBuilder"></a>
-## type [CmdBuilder](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/wiringcmd/cmdbuilder.go#L86-L96>)
+## type [CmdBuilder](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/cmdbuilder/cmdbuilder.go#L86-L96>)
 
 A helper struct when a Blueprint application supports multiple different wiring specs. Makes it easy to choose which spec to compile. See the Blueprint example applications for usage
 
@@ -105,7 +105,7 @@ type CmdBuilder struct {
 ```
 
 <a name="NewCmdBuilder"></a>
-### func [NewCmdBuilder](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/wiringcmd/cmdbuilder.go#L116>)
+### func [NewCmdBuilder](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/cmdbuilder/cmdbuilder.go#L116>)
 
 ```go
 func NewCmdBuilder(applicationName string) *CmdBuilder
@@ -114,7 +114,7 @@ func NewCmdBuilder(applicationName string) *CmdBuilder
 
 
 <a name="CmdBuilder.Add"></a>
-### func \(\*CmdBuilder\) [Add](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/wiringcmd/cmdbuilder.go#L123>)
+### func \(\*CmdBuilder\) [Add](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/cmdbuilder/cmdbuilder.go#L123>)
 
 ```go
 func (b *CmdBuilder) Add(specs ...SpecOption)
@@ -123,7 +123,7 @@ func (b *CmdBuilder) Add(specs ...SpecOption)
 
 
 <a name="CmdBuilder.Build"></a>
-### func \(\*CmdBuilder\) [Build](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/wiringcmd/cmdbuilder.go#L173>)
+### func \(\*CmdBuilder\) [Build](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/cmdbuilder/cmdbuilder.go#L173>)
 
 ```go
 func (b *CmdBuilder) Build() error
@@ -132,7 +132,7 @@ func (b *CmdBuilder) Build() error
 
 
 <a name="CmdBuilder.List"></a>
-### func \(\*CmdBuilder\) [List](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/wiringcmd/cmdbuilder.go#L165>)
+### func \(\*CmdBuilder\) [List](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/cmdbuilder/cmdbuilder.go#L165>)
 
 ```go
 func (builder *CmdBuilder) List() string
@@ -141,7 +141,7 @@ func (builder *CmdBuilder) List() string
 Returns a list of configured wiring specs
 
 <a name="CmdBuilder.ParseArgs"></a>
-### func \(\*CmdBuilder\) [ParseArgs](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/wiringcmd/cmdbuilder.go#L129>)
+### func \(\*CmdBuilder\) [ParseArgs](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/cmdbuilder/cmdbuilder.go#L129>)
 
 ```go
 func (b *CmdBuilder) ParseArgs()
@@ -150,7 +150,7 @@ func (b *CmdBuilder) ParseArgs()
 
 
 <a name="CmdBuilder.ValidateArgs"></a>
-### func \(\*CmdBuilder\) [ValidateArgs](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/wiringcmd/cmdbuilder.go#L141>)
+### func \(\*CmdBuilder\) [ValidateArgs](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/cmdbuilder/cmdbuilder.go#L141>)
 
 ```go
 func (b *CmdBuilder) ValidateArgs() error
@@ -159,7 +159,7 @@ func (b *CmdBuilder) ValidateArgs() error
 
 
 <a name="SpecOption"></a>
-## type [SpecOption](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/wiringcmd/cmdbuilder.go#L77-L81>)
+## type [SpecOption](<https://github.com/blueprint-uservices/blueprint/blob/main/plugins/cmdbuilder/cmdbuilder.go#L77-L81>)
 
 A wiring spec option used by [CmdBuilder](<#CmdBuilder>). When running the program, this wiring spec can be selected by specifying its \[Name\] with the \-w flag, e.g.
